requirements.txt
requests>=2.31.0
feedparser>=6.0.10
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
pyyaml>=6.0

